fix: resolve phpmd violations for pcpignore matching
- Extract split_anchored_entries(), is_file_in_ignored_directory() and
  is_file_ignored() into a new Ignore_Matcher class to keep
  Plugin_Request_Utility under the TooManyPublicMethods threshold.
- Split PCP_Ignore_Utility::get_exclusions() into smaller
  resolve_plugin_directory(), read_ignore_file_lines() and
  parse_exclusions() helpers to reduce NPathComplexity.
