+++ includes/Utilities/PCP_Ignore_Utility.php
@@ -57,11 +57,36 @@ class PCP_Ignore_Utility {
 	public static function get_exclusions( $plugin_path ) {
 		self::$warning = '';
 
-		$exclusions = array(
+		$empty_exclusions = array(
 			'directories' => array(),
 			'files'       => array(),
 		);
 
+		$plugin_directory = self::resolve_plugin_directory( $plugin_path );
+
+		if ( null === $plugin_directory ) {
+			return $empty_exclusions;
+		}
+
+		$lines = self::read_ignore_file_lines( trailingslashit( $plugin_directory ) . '.pcpignore' );
+
+		if ( null === $lines ) {
+			return $empty_exclusions;
+		}
+
+		return self::parse_exclusions( $lines );
+	}
+
+	/**
+	 * Resolves the plugin directory that may hold a `.pcpignore` file.
+	 *
+	 * @since n.e.x.t
+	 *
+	 * @param string $plugin_path Plugin directory or main plugin file path.
+	 * @return string|null The normalized, untrailingslashed plugin directory, or null
+	 *                      if the plugin is a single-file plugin, which is not supported.
+	 */
+	private static function resolve_plugin_directory( $plugin_path ) {
 		$plugin_directory = is_dir( $plugin_path ) ? $plugin_path : dirname( $plugin_path );
 		$plugin_directory = untrailingslashit( wp_normalize_path( $plugin_directory ) );
 
@@ -71,13 +96,27 @@ class PCP_Ignore_Utility {
 		// meant for one plugin to every other plugin scanned from that shared
 		// location, so .pcpignore is not supported for single-file plugins.
 		if ( untrailingslashit( wp_normalize_path( WP_PLUGIN_DIR ) ) === $plugin_directory ) {
-			return $exclusions;
+			return null;
 		}
 
-		$ignore_file = trailingslashit( $plugin_directory ) . '.pcpignore';
+		return $plugin_directory;
+	}
 
+	/**
+	 * Reads and returns the non-empty lines of a `.pcpignore` file.
+	 *
+	 * Records a warning, retrievable via {@see self::get_warning()}, if the
+	 * file exists but could not be read or parsed.
+	 *
+	 * @since n.e.x.t
+	 *
+	 * @param string $ignore_file Absolute path to the `.pcpignore` file.
+	 * @return array|null The file lines, or null if the file does not exist,
+	 *                     could not be read, or could not be parsed.
+	 */
+	private static function read_ignore_file_lines( $ignore_file ) {
 		if ( ! file_exists( $ignore_file ) ) {
-			return $exclusions;
+			return null;
 		}
 
 		if ( ! is_readable( $ignore_file ) ) {
@@ -87,7 +126,7 @@ class PCP_Ignore_Utility {
 				$ignore_file
 			);
 
-			return $exclusions;
+			return null;
 		}
 
 		$lines = file( $ignore_file, FILE_IGNORE_NEW_LINES );
@@ -99,8 +138,25 @@ class PCP_Ignore_Utility {
 				$ignore_file
 			);
 
-			return $exclusions;
+			return null;
 		}
+
+		return $lines;
+	}
+
+	/**
+	 * Parses `.pcpignore` file lines into directory and file exclusions.
+	 *
+	 * @since n.e.x.t
+	 *
+	 * @param array $lines The `.pcpignore` file lines.
+	 * @return array{directories: array, files: array} Custom exclusions, anchored to the plugin root.
+	 */
+	private static function parse_exclusions( array $lines ) {
+		$exclusions = array(
+			'directories' => array(),
+			'files'       => array(),
+		);
 
 		foreach ( $lines as $line ) {
 			$line = trim( $line );
